+++ wow_world_base/src/inner/shared/faction_flag_vanilla_tbc.rs
@@ -319,3 +319,74 @@ impl std::ops::BitXorAssign for FactionFlag {
     }
 }
 
+impl From<u8> for FactionFlag {
+    fn from(value: u8) -> Self {
+        Self::new(value)
+    }
+}
+
+impl TryFrom<u16> for FactionFlag {
+    type Error = u16;
+    fn try_from(value: u16) -> Result<Self, Self::Error> {
+        let a = TryInto::<u8>::try_into(value).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+
+impl TryFrom<u32> for FactionFlag {
+    type Error = u32;
+    fn try_from(value: u32) -> Result<Self, Self::Error> {
+        let a = TryInto::<u8>::try_into(value).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+
+impl TryFrom<u64> for FactionFlag {
+    type Error = u64;
+    fn try_from(value: u64) -> Result<Self, Self::Error> {
+        let a = TryInto::<u8>::try_into(value).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+
+impl From<i8> for FactionFlag {
+    fn from(value: i8) -> Self {
+        Self::new(u8::from_le_bytes(value.to_le_bytes()))
+    }
+}
+
+impl TryFrom<i16> for FactionFlag {
+    type Error = i16;
+    fn try_from(value: i16) -> Result<Self, Self::Error> {
+        let v = u16::from_le_bytes(value.to_le_bytes());
+        let a = TryInto::<u8>::try_into(v).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+
+impl TryFrom<i32> for FactionFlag {
+    type Error = i32;
+    fn try_from(value: i32) -> Result<Self, Self::Error> {
+        let v = u32::from_le_bytes(value.to_le_bytes());
+        let a = TryInto::<u8>::try_into(v).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+
+impl TryFrom<i64> for FactionFlag {
+    type Error = i64;
+    fn try_from(value: i64) -> Result<Self, Self::Error> {
+        let v = u64::from_le_bytes(value.to_le_bytes());
+        let a = TryInto::<u8>::try_into(v).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+
+impl TryFrom<usize> for FactionFlag {
+    type Error = usize;
+    fn try_from(value: usize) -> Result<Self, Self::Error> {
+        let a = TryInto::<u8>::try_into(value).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+