+++ wow_world_base/src/inner/vanilla/allowed_race.rs
@@ -380,3 +380,70 @@ impl std::ops::BitXorAssign for AllowedRace {
     }
 }
 
+impl From<u32> for AllowedRace {
+    fn from(value: u32) -> Self {
+        Self::new(value)
+    }
+}
+
+impl From<u8> for AllowedRace {
+    fn from(value: u8) -> Self {
+        Self::new(value.into())
+    }
+}
+
+impl From<u16> for AllowedRace {
+    fn from(value: u16) -> Self {
+        Self::new(value.into())
+    }
+}
+
+impl TryFrom<u64> for AllowedRace {
+    type Error = u64;
+    fn try_from(value: u64) -> Result<Self, Self::Error> {
+        let a = TryInto::<u32>::try_into(value).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+
+impl TryFrom<i8> for AllowedRace {
+    type Error = i8;
+    fn try_from(value: i8) -> Result<Self, Self::Error> {
+        let v = u8::from_le_bytes(value.to_le_bytes());
+        let a = TryInto::<u32>::try_into(v).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+
+impl TryFrom<i16> for AllowedRace {
+    type Error = i16;
+    fn try_from(value: i16) -> Result<Self, Self::Error> {
+        let v = u16::from_le_bytes(value.to_le_bytes());
+        let a = TryInto::<u32>::try_into(v).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+
+impl From<i32> for AllowedRace {
+    fn from(value: i32) -> Self {
+        Self::new(u32::from_le_bytes(value.to_le_bytes()))
+    }
+}
+
+impl TryFrom<i64> for AllowedRace {
+    type Error = i64;
+    fn try_from(value: i64) -> Result<Self, Self::Error> {
+        let v = u64::from_le_bytes(value.to_le_bytes());
+        let a = TryInto::<u32>::try_into(v).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+
+impl TryFrom<usize> for AllowedRace {
+    type Error = usize;
+    fn try_from(value: usize) -> Result<Self, Self::Error> {
+        let a = TryInto::<u32>::try_into(value).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+