+++ wow_world_base/src/inner/wrath/movement_flags.rs
@@ -1526,3 +1526,71 @@ impl std::ops::BitXorAssign for MovementFlags {
     }
 }
 
+impl From<u64> for MovementFlags {
+    fn from(value: u64) -> Self {
+        Self::new(value)
+    }
+}
+
+impl From<u8> for MovementFlags {
+    fn from(value: u8) -> Self {
+        Self::new(value.into())
+    }
+}
+
+impl From<u16> for MovementFlags {
+    fn from(value: u16) -> Self {
+        Self::new(value.into())
+    }
+}
+
+impl From<u32> for MovementFlags {
+    fn from(value: u32) -> Self {
+        Self::new(value.into())
+    }
+}
+
+impl TryFrom<i8> for MovementFlags {
+    type Error = i8;
+    fn try_from(value: i8) -> Result<Self, Self::Error> {
+        let v = u8::from_le_bytes(value.to_le_bytes());
+        let a = TryInto::<u64>::try_into(v).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+
+impl TryFrom<i16> for MovementFlags {
+    type Error = i16;
+    fn try_from(value: i16) -> Result<Self, Self::Error> {
+        let v = u16::from_le_bytes(value.to_le_bytes());
+        let a = TryInto::<u64>::try_into(v).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+
+impl TryFrom<i32> for MovementFlags {
+    type Error = i32;
+    fn try_from(value: i32) -> Result<Self, Self::Error> {
+        let v = u32::from_le_bytes(value.to_le_bytes());
+        let a = TryInto::<u64>::try_into(v).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+
+impl TryFrom<i64> for MovementFlags {
+    type Error = i64;
+    fn try_from(value: i64) -> Result<Self, Self::Error> {
+        let v = u64::from_le_bytes(value.to_le_bytes());
+        let a = TryInto::<u64>::try_into(v).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+
+impl TryFrom<usize> for MovementFlags {
+    type Error = usize;
+    fn try_from(value: usize) -> Result<Self, Self::Error> {
+        let a = TryInto::<u64>::try_into(value).ok().ok_or(value)?;
+        Ok(Self::new(a))
+    }
+}
+